% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MortgageP}
\alias{MortgageP}
\title{mortgage package}
\description{
The package has one function that returns an amortization table that lists monthly payment on a mortgageloan based on a loan amount, duration of loan in years, and annual interest rate. The table will provide the breakdown of the payment by how much is allocated to interest and the loan principal. It will also list the loan balance for each month.
}
\author{
Meryem El Rhazouani
}
