% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MortgageP}
\alias{MortgageP}
\title{mortgage package}
\description{
The package gives a table that lists each regular payment on a mortgage based on a loan amount, annual interest rate, and duration of the laon.
}
\author{
Meryem El Rhazouani
}
